Add spouse variables and update main script
This commit introduces the `09_spouse.r` script, which identifies spousal information by preparing and merging spousal variables into the existing CFPS dataset. Additionally, the `00_main.r` script is updated to source the new spouse script and provide a message indicating the addition of spouse variables. Codes for saving the data is moved from `08_merge.r` to the new script.